reg/requirements.txt
selenium>=4.0.0
undetected-chromedriver>=3.5.0
requests>=2.28.0
PyYAML>=6.0
